Migrate common accessibility functionality to common module
